[cleanup] Various cleanups (#18)
* [account] Cleanup account to have comments, reuse code

* [refactor] Rename Account0x1 to AccountOne to match other SDKs

* [bcs] Uleb128 only supports uint32

* [refactor] Misc documentation and some cleanup

* [ed25519] Add support for ToHex and FromHex for keys

+++ typetag.go
@@ -32,7 +32,7 @@ type TypeTag struct {
 }
 
 func (tt *TypeTag) MarshalBCS(bcs *Serializer) {
-	bcs.Uleb128(uint64(tt.Value.GetType()))
+	bcs.Uleb128(uint32(tt.Value.GetType()))
 	tt.Value.MarshalBCS(bcs)
 
 }
